requirements.txt
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
python-dotenv
mongoengine
flask-mongoengine
# psycopg2-binary
# Flask-PyMongo
# Flask-MySQL
# postgres
# flask_postgres
#https://python.plainenglish.io/some-useful-python-flask-plugins-3ac57faaca7e
# Flask-SQLAlchemy
# Flask-WTF
# Flask-Uploads
# Flask-Debugtoolbar
# Flask-Login
# Flask-Admin
# Flask-Cors
# Flask-Migrate
# peewee
# wtf-peewee
# flask-peewee
#https://nickjanetakis.com/blog/15-useful-flask-extensions-and-libraries-that-i-use-in-every-project
# alembic
#https://gitee.com/labike/awesome-flask
# Flask-Restless
# Flask-User
# Flask-SuperAdmin
# Flask-Alembic
